Address PR review feedback on draft helpers.
Tighten summarize characterization, cover legacy isPaid drafts, share createSafeDate with submit path, and assert date on round-trip.
